fix: remove Ekubo dependency for stabilizer (#11)
This commit removes Ekubo as a dependency due to issues encountered 
when trying to publish on scarbs.xyz. There is a requirement that 
packages have a version specified, but Ekubo is not published on 
scarbs.xyz. Therefore, functions that rely on Ekubo's types have been 
commented out.

Additionally, this commit fixes an omission to include the stabilizer interface 
for compilation as well as for export.